Add files via upload
* Isolated Redis Functions
* Added Faux Admin Login to see reports ( TODO: Secure Login )
* Refactored some server side code
