+++ public/main.html
@@ -8,6 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     
     <link href="//maxcdn.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css" rel="stylesheet" id="bootstrap-css">
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
     <link rel="stylesheet" href="./assets/css/main.css">
     
   </head>
@@ -59,13 +60,13 @@
 								<div class="form-group">
 									<label class="control-label col-sm-2" for="name">Name:</label>
 									<div class="col-sm-10">          
-										<input type="text" class="form-control inputstl" id="name" placeholder="Enter Your Name" name="name" required>
+										<input type="text" class="form-control inputstl" id="name" placeholder="Enter Your Name" name="name" maxlength="32" required>
 									</div>
 								</div>
 								<div class="form-group">
 									<label class="control-label col-sm-2" for="location">Location</label>
 									<div class="col-sm-10">          
-										<input type="text" class="form-control inputstl" id="location" placeholder="Enter location" name="location" required>
+										<input type="text" class="form-control inputstl" id="location" placeholder="Enter location" name="location" maxlength="32" required>
 									</div>
 								</div>
 								<div class="form-group">
@@ -77,25 +78,25 @@
 								<div class="form-group">
 									<label class="control-label col-sm-2" for="duckCount">Duck Count</label>
 									<div class="col-sm-10">          
-										<input type="number"  class="form-control inputstl" id="duckCount" placeholder="How many ducks were fed" name="duckCount" min="1" required>
+										<input type="number"  class="form-control inputstl" id="duckCount" placeholder="How many ducks were fed" name="duckCount" min="1" max="100" required>
 									</div>
 								</div>
 								<div class="form-group">
 									<label class="control-label col-sm-2" for="duckFood">Food Type</label>
 									<div class="col-sm-10">          
-										<input type="text" class="form-control inputstl" id="duckFood" placeholder="What were the ducks fed" name="duckFood" required>
+										<input type="text" class="form-control inputstl" id="duckFood" placeholder="What were the ducks fed" name="duckFood" maxlength="32" required>
 									</div>
 								</div>
 								<div class="form-group">
 									<label class="control-label col-sm-2" for="foodCount">Food Quantity</label>
 									<div class="col-sm-10">          
-										<input type="text" class="form-control inputstl" id="foodCount" placeholder="Quantity of feed" name="foodCount" required>
+										<input type="text" class="form-control inputstl" id="foodCount" placeholder="Quantity of feed" name="foodCount" maxlength="32" required>
 									</div>
 								</div>
 								<div class="form-group">
 									<label class="control-label col-sm-2" for="comment">Comments:</label>
 									<div class="col-sm-10">
-									<textarea class="form-control inputstl" rows="4" id="comment" name="comment"></textarea>
+									<textarea class="form-control inputstl" rows="4" id="comment" name="comment" maxlength="255"></textarea>
 									</div>
 								</div>
 								<div class="form-group">
@@ -127,23 +128,24 @@
 						<span aria-hidden="true">&times;</span>
 					</button>
 				</div>
-				<div class="modal-body mx-3">
-					<div class="md-form mb-5">
-						<i class="fas fa-envelope prefix grey-text"></i>
-						<input type="email" id="defaultForm-email" class="form-control validate">
-						<label data-error="wrong" data-success="right" for="defaultForm-email">Your email</label>
-					</div>
+					<form action="http://localhost:8080/login" method="post">
+						<div class="modal-body mx-3">
+							<div class="md-form mb-5">
+								<i class="fas fa-envelope prefix grey-text"></i>
+								<input type="email" id="defaultForm-email" class="form-control validate">
+								<label data-error="wrong" data-success="right" for="defaultForm-email">Your email</label>
+							</div>
 
-					<div class="md-form mb-4">
-						<i class="fas fa-lock prefix grey-text"></i>
-						<input type="password" id="defaultForm-pass" class="form-control validate">
-						<label data-error="wrong" data-success="right" for="defaultForm-pass">Your password</label>
+							<div class="md-form mb-4">
+								<i class="fas fa-lock prefix grey-text"></i>
+								<input type="password" id="defaultForm-pass" class="form-control validate">
+								<label data-error="wrong" data-success="right" for="defaultForm-pass">Your password</label>
+							</div>
+						</div>
+					<div class="modal-footer d-flex justify-content-center">
+						<button type="submit" class="btn btn-default">Login</button>
 					</div>
-
-				</div>
-				<div class="modal-footer d-flex justify-content-center">
-					<button class="btn btn-default">Login</button>
-				</div>
+				</form>
 			</div>
 		</div>
 	</div>
