+++ public/success.html
@@ -1,1 +1,1 @@
-<h1>Thanks for the submission!</h1>
+<h1>This is the success Page.. thanks for your data.. Don't forget to feed the ducks tomorrow.</h1>
