+++ servJS/servUtil.js
@@ -1,0 +1,15 @@
+'use strict';
+
+
+
+function stringCheck(str){
+	
+	if(!typeof str == 'string'){
+		return;
+	}
+	
+	let str = str.trim();
+	return str;
+}
+
+module.exports = {};
